unify si_ssim & ssim
